rsoi-2022-lab2-microservices-malshevskikh-master/rsoi-2022-lab2-microservices-malshevskikh-master/source_v1/TicketService/requirements.txt
asgiref==3.5.2
Babel==2.8.0
backports.zoneinfo==0.2.1
certifi==2020.6.20
chardet==3.0.4
charset-normalizer==2.1.1
click==8.0.3
cloudpickle==2.0.0
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
dask==2021.10.0
DateTime==4.7
decorator==4.4.2
defusedxml==0.7.1
distributed==2021.10.0
Django==4.1.1
djangorestframework==3.14.0
docopt==0.6.2
fonttools==4.38.0
fsspec==2021.10.1
HeapDict==1.0.1
httpie==3.2.1
idna==2.10
Jinja2==3.0.2
kiwisolver==1.4.4
locket==0.2.1
MarkupSafe==2.0.1
matplotlib==3.6.2
msgpack==1.0.2
multidict==6.0.2
numpy==1.19.4
packaging==21.0
partd==1.2.0
pi==0.1.2
Pillow==9.3.0
pipreqs==0.4.11
psutil==5.8.0
psycopg2-binary==2.9.3
Pygments==2.13.0
pyparsing==3.0.1


PySocks==1.7.1
python-dateutil==2.8.2
pytz==2020.1
PyYAML==6.0
qtwidgets==0.17
requests==2.24.0
requests-toolbelt==0.9.1
rich==12.6.0

shiboken2==5.15.1
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.4.3
tblib==1.7.0
tkcalendar==1.6.1
toolz==0.11.1
tornado==6.1
typing_extensions==4.3.0
urllib3==1.25.10
views==0.3
yarg==0.1.9
zict==2.0.0
zope.interface==5.5.2
